Move open in editor out of errors
