+++ packages/reporter/src/files/file-model.ts
@@ -1,0 +1,7 @@
+export interface FileDetails {
+    absoluteFile: string
+    column: number
+    line: number
+    originalFile: string
+    relativeFile: string
+}
